+++ src/module/vue/components/impact/impact-checkbox.vue
